Match colored light intensity better to original during .lit loading. Change bestcolor distortion to favor value over hue. Load/savegame view angle fix thanks to mh. Fix window getting stuck when moved too far off-screen.  Thanks to dreadlorde for bug report.

+++ client.h
@@ -47,7 +47,7 @@ typedef struct
 typedef struct
 {
 	int		destcolor[3];
-	int		percent;		// 0-256
+	float		percent;		// 0-256  //qbism - was int.  float for higher precision (thanks mh)
 } cshift_t;
 
 //qbism- from engoo... leilei - shadowhack
